Fixed Ocean Bug
When there are no drops and the user cancels a "add drop" request
the map gets redirected to the ocean of a universal default position,
bug has been fixed by attempting to grab user location... However
the alternative has been applied to where the map redirection stays
where the user attempted to drop the drop. So that's nice.

Please enter the commit message for your changes. Lines starting